requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
numpy>=1.9.3
torchvision>=0.15.2
torch>=2.1.0
tqdm>=4.65.0
pillow>=10.2.0
transformers>=4.32.1
requests>=2.28.1
matplotlib>=3.8.3
